Refined ip prompt for macOS
The old prompt was giving me multiple IP addresses whenever I started up
a VMware virtual machine. The new way calls like a bunch of functions to
do the same thing, but I don't notice it to be any slower. It also
actually gives the one and only correct IP address as reported by the
'Remote Login' section of the 'Sharing' option in macOS System
Preferences.